split home to componenets
